requirements.txt
python-dotenv==1.0.1
Flask==3.0.2
discord.py==2.4.0
